Refactoring of Asset Batches, Window logic and changed frame delta time to long

+++ SandAndStonesEngine/Assets/GameAsset.cs
@@ -31,7 +31,7 @@ namespace SandAndStonesEngine.Assets
             base.SetParam(param);
         }
 
-        public override void Update(double delta)
+        public override void Update(long delta)
         {
             base.Update(delta);
         }
